Jade: Minor improvements to AI code

+++ Assets/Scripts/DataTypes/Jade/AI/Vars/AI_VarType.cs
@@ -4,7 +4,7 @@
 		Unknown,
 		Bool,
 		Int,
-		UInt,
+		Float,
 		Vector,
 		String,
 		PointerRef,
